% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{loadData}
\usage{
loadData(path, sample_pct)
}
\arguments{
\item{path}{System path to raw data file}

\item{sample_pct}{Random sample to take of data [0,1]}
}
\value{
data frame of formatted data
}
\description{
Loads and formats source data into data needed in application
}
\details{
Fuller description to be added when SQL language is integrated in here (for now we just parse a random data file)
}
\examples{
\dontrun{
path <- c('~/CHR_Reference_Data/FullDataSet_AllYearsCombined.txt')
sample_pct <- 0.1
loadData(path,sample_pct)
}
}

